need make refactor with styles
