+++ src/screens/FamilyInfoOverviewScreen.js
@@ -1,13 +1,15 @@
 import { View, StyleSheet, ScrollView } from "react-native";
-import { FamilyInfoTop } from "../components/FamilyInfoPage/FamilyInfoTop";
-import { FamilyInfoItem } from "../components/FamilyInfoPage/FamilyInfoItem";
+import { FamilyInfoTop } from "../components/familyInfoPage/FamilyInfoTop";
+import { PersonalInfoItem } from "../components/familyInfoPage/PersonalInfoItem";
+import { EmploymentInfoItem } from "../components/familyInfoPage/EmploymentInfoItem";
 
 export const FamilyInfoOverviewScreen = () => {
   return (
     <View style={styles.wrapper}>
       <ScrollView stickyHeaderIndices={[0]}>
         <FamilyInfoTop />
-        <FamilyInfoItem />
+        <PersonalInfoItem />
+        <EmploymentInfoItem/>
       </ScrollView>
     </View>
   );
